# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Connect | Server",
    "version": "2.0.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/bases": {
      "post": {
        "operationId": "createBase",
        "summary": "Create a new base",
        "tags": [
          "Base"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "document": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number",
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": "number",
                    "minimum": -180,
                    "maximum": 180
                  },
                  "unitId": {
                    "type": "string"
                  },
                  "phones": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "number": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "isWhatsapp": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "number"
                      ]
                    }
                  }
                },
                "required": [
                  "name",
                  "latitude",
                  "longitude",
                  "unitId",
                  "phones"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getBases",
        "summary": "Get bases",
        "tags": [
          "Base"
        ],
        "description": "Get bases",
        "parameters": [
          {
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "default": 10,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "perPage",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "name",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "query",
            "name": "unitId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "query",
            "name": "companyId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "query",
            "name": "companyGroupId",
            "required": false
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "cuid",
                            "pattern": "^[cC][^\\s-]{8,}$"
                          },
                          "name": {
                            "type": "string"
                          },
                          "unit": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "cuid",
                                "pattern": "^[cC][^\\s-]{8,}$"
                              },
                              "name": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name"
                            ],
                            "additionalProperties": false
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "unit",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "hasPreviousPage": {
                          "type": "boolean"
                        },
                        "currentPage": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "total",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage",
                        "currentPage"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/companies/{companyId}": {
      "put": {
        "operationId": "updateCompany",
        "summary": "Update company",
        "tags": [
          "Company"
        ],
        "description": "Update company",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "document": {
                    "description": "Brazilian CNPJ",
                    "type": "string",
                    "minLength": 14,
                    "maxLength": 14
                  }
                },
                "required": [
                  "name",
                  "document"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "companyId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteCompany",
        "summary": "Delete company",
        "tags": [
          "Company"
        ],
        "description": "Delete company",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "companyId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "nullable": true,
                  "enum": [
                    null
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getCompany",
        "summary": "Get company",
        "tags": [
          "Company"
        ],
        "description": "Get company",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "cuid",
              "pattern": "^[cC][^\\s-]{8,}$"
            },
            "in": "path",
            "name": "companyId",
            "required": true
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "cuid",
                          "pattern": "^[cC][^\\s-]{8,}$"
                        },
                        "name": {
                          "type": "string"
                        },
                        "document": {
                          "type": "string"
                        },
                        "companyGroupId": {
                          "type": "string",
                          "format": "cuid",
                          "pattern": "^[cC][^\\s-]{8,}$"
                        },
                        "unitsCount": {
                          "type": "number"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "companyGroup": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "cuid",
                              "pattern": "^[cC][^\\s-]{8,}$"
                            },
                            "name": {
                              "type": "string"
                            },
                            "document": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "document"
                          ],
                          "additionalProperties": false
                        },
                        "phones": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "cuid",
                                "pattern": "^[cC][^\\s-]{8,}$"
                              },
                              "name": {
                                "nullable": true,
                                "type": "string"
                              },
                              "isWhatsapp": {
                                "type": "boolean"
                              },
                              "number": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "id",
                              "name",
                              "isWhatsapp",
                              "number",
                              "createdAt"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "document",
                        "companyGroupId",
                        "unitsCount",
                        "createdAt",
                        "updatedAt",
                        "companyGroup",
                        "phones"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}